feat: add equality checks for error recovery attempter
This is needed for the coalescing feature so that it can be checked whether a recovery attempter is equal to another or not. As we can not check the internals of the recovery block, we are making a best guess here in checking whether the localised recovery options are equal on both recovery attempters.